Eliminate duplicate css files using hashing.
Parsing css files is by far the most time consuming part of the aggregator. Over 90% of the wall time is used parsing css into ASTs. In a test project with around 250 css files, with many duplicates, parsing takes around 14 seconds on s 2,5GHz Core i5 M processer, while all the other tasks combined take less than two seconds. Therefore we want to do as little parsing as possible.

The changes in this commit compares the xxHash of each file with a set of hashes of already processed files. Duplicates are then eliminated before they are handed over to the css parser. This results in a factor 8 speedup of the aggregation function on the test project, which now completes in around two seconds.